+++ packages/dash-table/src/dash-table/dash/Sanitizer.ts
@@ -16,7 +16,8 @@ import {
     ExportFormat,
     ExportHeaders,
     IFilterAction,
-    FilterLogicalOperator
+    FilterLogicalOperator,
+    SelectedCells
 } from 'dash-table/components/Table/props';
 import headerRows from 'dash-table/derived/header/headerRows';
 import resolveFlag from 'dash-table/derived/cell/resolveFlag';
@@ -33,6 +34,7 @@ const D3_DEFAULT_LOCALE: INumberLocale = {
 
 const DEFAULT_NULLY = '';
 const DEFAULT_SPECIFIER = '';
+const NULL_SELECTED_CELLS: SelectedCells = [];
 
 const data2number = (data?: any) => +data || 0;
 
@@ -99,7 +101,16 @@ export default class Sanitizer {
             headerFormat = ExportHeaders.Ids;
         }
 
+        const active_cell = props.cell_selectable ?
+            props.active_cell :
+            undefined;
+
+        const selected_cells = props.cell_selectable ?
+            props.selected_cells :
+            NULL_SELECTED_CELLS;
+
         return R.merge(props, {
+            active_cell,
             columns,
             data,
             export_headers: headerFormat,
@@ -108,6 +119,7 @@ export default class Sanitizer {
             fixed_rows: getFixedRows(props.fixed_rows, columns, props.filter_action),
             loading_state: dataLoading(props.loading_state),
             locale_format,
+            selected_cells,
             visibleColumns
         });
     }