+++ packages/dash-table/tests/selenium/test_navigation.py
@@ -185,3 +185,21 @@ def test_navg004_keyboard_between_md_and_standard_cells(test, props):
         test.send_keys(Keys.ARROW_RIGHT)
         test.send_keys(Keys.ARROW_DOWN)
         assert target.cell(i, i).is_focused()
+
+
+@pytest.mark.parametrize("cell_selectable", [True, False])
+def test_navg005_unselectable_cells(test, cell_selectable):
+    app = dash.Dash(__name__)
+    app.layout = DataTable(
+        id="table",
+        columns=[dict(id="a", name="a"), dict(id="b", name="b")],
+        data=[dict(a=0, b=0), dict(a=1, b=2)],
+        cell_selectable=cell_selectable,
+    )
+
+    test.start_server(app)
+
+    target = test.table("table")
+    target.cell(0, "a").click()
+
+    assert target.cell(0, "a").is_selected() == cell_selectable